requirements/monitoring_extras.txt
google-api-python-client==1.10.0
google-cloud-logging==1.15.1